Feature: Inventory management for products added to the stripecheckout service

+++ Application/Interfaces/Services/GenericService/IGenericService.cs
@@ -9,7 +9,7 @@ namespace BusinessLogic.Interfaces.Services.GenericService
 {
     public interface IGenericService<T> where T : class
     {
-        Task<T> Get(int id);
+        Task<T> Get(long id);
         Task<IEnumerable<T>> Get();
         Task<T> post(T Entity);
         Task Put(T Entity);
